Move slave registry into framework package
