+++ contrib/mesos/pkg/scheduler/components/framework/driver_mock.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package mock
+package framework
 
 import (
 	mesos "github.com/mesos/mesos-go/mesosproto"
